requirements.txt
be-great==0.0.7
gensim==4.3.2
kiwisolver==1.4.5
kneed==0.8.5
matplotlib==3.8.2
nltk==3.8.1
numpy==1.26.4
pandas==2.2.0
pyarrow==15.0.0
python-snappy==0.7.2
PyYAML==6.0.1
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
tabgan==2.2.1
zstandard==0.22.0
tabulate==0.9.0